fix(kg): remediate PR #1114 review findings (issues #818/#816)
1. Custom label validation: add EdgeKindError + EdgeKind::custom()
   checked constructor that rejects empty labels and colon-containing
   labels to prevent double-prefix on-disk tags. Fix from_tag() to
   return None for malformed "custom:"-prefixed suffixes (empty or
   with ':'). Add EdgeKindError to public symgraph exports.

2. Warm-boot test integrity: test_custom_edge_survives_warm_boot
   already used all_edges() (not callees_of) — confirmed adequate.
   Add doc notes on callers_of/callees_of in traverse.rs documenting
   the CallsFunction-only design so callers use neighbors_by_edge or
   all_edges for non-call edge kinds.

3. fact_hash_str provenance: PRE-EXISTING — function existed on main
   with the same DefaultHasher implementation; this PR only moved it
   during the contracts split. Fix the misleading doc comment to
   remove false cross-version stability claim; add caveat noting
   DefaultHasher is NOT stable across Rust versions/process runs.
   Algorithm unchanged to avoid invalidating persisted entity IDs.

4. Trivial nits: static_tag Custom arm panic → unreachable!;
   edge_kind_union_coverage now constructs all 30 variants and asserts
   each tag() is non-empty and round-trips via from_tag(); duplicate
   structured-field + interpolated count in load_from_corpus warn
   reduced to structured field only.

5. File split: edge_kind.rs was 573 lines after additions; test module
   extracted to edge_kind_tests.rs (204 lines) bringing edge_kind.rs
   to 380 lines — both under the 500-line cap. Wire via
   #[cfg(test)] #[path = "edge_kind_tests.rs"] mod tests.
